requirements.txt
gTTS
playsound
playsound3
sounddevice
vosk
opencv-python
ultralytics
torch
numpy
mtcnn
keras-facenet
requests
pytesseract
Pillow
easyocr
pygame
mutagen